Remove utc time of the working group meetings
With daylight saving time this would currently be at 7pm UTC. Having a wrong date is misleading while updating it twice a year or describing this would be too complicated. 
I thought about linking the date to a website converting it to the users local time but did not want to choose a one.