Clarify that anyone can participate in the project
Adds an "Everyone" role to try and clarify that you do not need to be a "contributor" to contribute. It's just a perhaps unfortunate choice for the name of the role we've chosen for folks that regularly participate in the Paketo projects. Everyone can contribute, those with the contributor role contribute a lot.

Resolves #84

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>